change answer 102-fibo

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,7 +8,7 @@
 void fibonacci(void)
 {
 	int i;
-	long long mult = 1;
+	unsigned long mult = 1;
 	int last = 0;
 
 	for (i = 0; i <= 50; i++)
@@ -18,7 +18,7 @@ void fibonacci(void)
 		current = mult;
 		mult = mult + last;
 		last = current;
-		printf("%1lld", mult);
+		printf("%1ld", mult);
 		putchar(',');
 		putchar(' ');
 	}
